refactor: 디렉터리 구조 변경, validator 스태틱으로 처리
