--- src/js/Model.js
@@ -1,34 +1,0 @@
-class LottoModel {
-  constructor() {
-    this._lotto = [];
-    //this._money = 0;
-    //this._count = 0;
-    this.detail = false;
-  }
-
-  get lootto() {
-    return this._lotto;
-  }
-
-  get money() {
-    return this._money;
-  }
-
-  get count() {
-    return this._count;
-  }
-
-  set lotto(newLotto) {
-    this._lotto = newLotto;
-  }
-
-  set money(newMoney) {
-    this._money = newMoney;
-  }
-
-  set count(newCount) {
-    this._count = newCount;
-  }
-}
-
-export default LottoModel;
