+++ src/js/lottos/constant.js
