+++ src/js/lottos/model.js
@@ -1,9 +1,7 @@
-import { INIT } from "./constant.js";
-
 class LottoModel {
   constructor() {
-    this._lottos = INIT.LOTTOS;
-    this._detail = INIT.DETAIL;
+    this._lottos = [];
+    this._detail = false;
   }
 
   get lottos() {
@@ -31,8 +29,8 @@ class LottoModel {
   }
 
   init() {
-    this._lottos = INIT.LOTTOS;
-    this._detail = INIT.DETAIL;
+    this._lottos = [];
+    this._detail = false;
   }
 }
 export default LottoModel;
