+++ src/js/lottos/template.js
